backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.32
pandas==2.1.3
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1
